Added unique_id for lights
I ran into needing to set a self._attr_unique_id for the lights as a voice assistant needed the alias function/voice name for it. Setting the unique_id seems to cause the centralite_desc.yaml to be ignored and the friendly name needs to be entered into the web UI for each of the lights.